# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PDF-API",
    "version": "1.0"
  },
  "paths": {
    "/api/Print/TestConnection": {
      "post": {
        "tags": [
          "Print"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionTestRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionTestRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionTestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Print": {
      "post": {
        "tags": [
          "Print"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "PrinterName": {
                    "type": "string"
                  },
                  "ServerHost": {
                    "type": "string"
                  },
                  "ServerPort": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "File": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "PrinterName": {
                  "style": "form"
                },
                "ServerHost": {
                  "style": "form"
                },
                "ServerPort": {
                  "style": "form"
                },
                "File": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConnectionTestRequest": {
        "type": "object",
        "properties": {
          "host": {
            "type": "string",
            "nullable": true
          },
          "port": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}